# src/specification/swagger.yaml
openapi: 3.0.0
info:
  title: Sport Booking Backend API
  version: 1.0.0
  description: API документация для Sport Booking Backend
servers:
  - url: https://api.walknplay.ru
    description: Production server
  - url: http://localhost:3001
    description: Development server
...
paths:
  /auth/signin:
    post:
      summary: Вход пользователя по email и паролю
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Успешный вход (cookie auth_token устанавливается автоматически)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: ID пользователя
                    example: 550e8400-e29b-41d4-a716-446655440000
                  email_verified:
                    type: string
                    enum: [VERIFIED, NOT_VERIFIED]
                    description: Статус верификации email
                    example: NOT_VERIFIED
        '400':
          description: Ошибка валидации или неверные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signup:
    post:
      summary: Регистрация нового пользователя
      description: |
        Создает нового пользователя и отправляет письмо для подтверждения email.
        После подтверждения email через ссылку в письме пользователь будет автоматически авторизован.
        Если email не требует подтверждения, cookie auth_token устанавливается сразу.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
                role:
                  type: string
                  example: USER
                name:
                  type: string
                  example: Иван Иванов
                phone:
                  type: string
                  example: '+79001234567'
                date_of_birth:
                  type: string
                  format: date
                  example: '1990-01-15'
      responses:
        '200':
          description: Успешная регистрация (письмо подтверждения отправлено на email)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: ID пользователя
                    example: 550e8400-e29b-41d4-a716-446655440000
                  email_verified:
                    type: string
                    enum: [VERIFIED, NOT_VERIFIED]
                    description: Статус верификации email
                    example: NOT_VERIFIED
        '400':
          description: Ошибка валидации или пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/confirm:
    get:
      summary: Подтверждение email пользователя
      description: |
        Этот эндпоинт вызывается при переходе по ссылке из письма подтверждения.
        После успешного подтверждения устанавливается cookie auth_token и возвращаются данные пользователя.
      tags:
        - Authentication
      parameters:
        - in: query
          name: access_token
          required: true
          schema:
            type: string
          description: Токен доступа из письма подтверждения
      responses:
        '200':
          description: Успешное подтверждение email (cookie auth_token устанавливается автоматически)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: ID пользователя
                    example: 550e8400-e29b-41d4-a716-446655440000
                  email_verified:
                    type: string
                    enum: [VERIFIED]
                    description: Статус верификации email (всегда VERIFIED после подтверждения)
                    example: VERIFIED
        '400':
          description: Невалидная ссылка подтверждения
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signout:
    post:
      summary: Выход пользователя
      tags:
        - Authentication
      responses:
        '200':
          description: Успешный выход
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Signed out successfully
        '400':
          description: Ошибка при выходе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/reset-password:
    post:
      summary: Запросить восстановление пароля
      description: |
        Отправляет письмо с ссылкой для восстановления пароля на указанный email.
        Ссылка ведет на страницу /new-password с токеном в параметрах.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
      responses:
        '200':
          description: Письмо с инструкциями отправлено
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password reset requested
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/new-password:
    post:
      summary: Установить новый пароль
      description: |
        Устанавливает новый пароль пользователя после перехода по ссылке восстановления.
        Требуется access_token из ссылки восстановления пароля.
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
                - confirmPassword
                - access_token
              properties:
                password:
                  type: string
                  format: password
                  example: newPassword123
                confirmPassword:
                  type: string
                  format: password
                  example: newPassword123
                access_token:
                  type: string
                  description: Токен восстановления пароля (reset token)
                refresh_token:
                  type: string
                  description: Не используется (оставлено для совместимости)
      responses:
        '200':
          description: Пароль успешно обновлен (cookie auth_token устанавливается автоматически)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Password updated successfully
                  id:
                    type: string
                    description: ID пользователя
                  email_verified:
                    type: string
                    enum: [VERIFIED, NOT_VERIFIED]
                    description: Статус верификации email
        '400':
          description: Ошибка валидации или недействительный токен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    get:
      summary: Получить профиль текущего пользователя (требуется cookie auth_token)
      tags:
        - Profile
      security:
        - CookieAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Профиль не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Обновить профиль пользователя (требуется cookie auth_token)
      tags:
        - Profile
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                name:
                  type: string
                phone:
                  type: string
                date_of_birth:
                  type: string
                  format: date
      responses:
        '200':
          description: Профиль успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Внутренняя ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: ID пользователя
        email:
          type: string
          description: Email пользователя
        email_verified:
          type: string
          enum: [VERIFIED, NOT_VERIFIED]
          description: Статус верификации email
        role:
          type: string
          description: Роль пользователя
        name:
          type: string
          description: Имя пользователя
        phone:
          type: string
          description: Телефон пользователя
        date_of_birth:
          type: string
          format: date
          description: Дата рождения пользователя
        campaign_id:
          type: string
          description: ID кампании (только для пользователей с ролью CAMPAIGN)
        registration_date:
          type: string
          format: date-time
          description: Дата регистрации
        created_at:
          type: string
          format: date-time
          description: Дата создания
        updated_at:
          type: string
          format: date-time
          description: Дата обновления
    Error:
      type: object
      properties:
        error:
          type: string
          description: Сообщение об ошибке
...
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT токен в httpOnly cookie (устанавливается автоматически при логине/регистрации, действителен 7 дней)